PS-40: Renamed SettingsContainer to SettingsContainers
