+++ frontend/src/App.js
@@ -4,7 +4,7 @@ import './App.css';
 import Home from "./home_page/Home";
 import LoginPage from "./login_page/LoginPage";
 import ListPage from "./list_page/ListPage";
-import SettingsPage from "./settings/SettingsContainer"
+import SettingsPage from "./settings/SettingsContainers"
 import useToken from './services/useToken';
 
 /**
